refactor: retire legacy scheduler protocol
