+++ src/domain/mod.rs
@@ -2,7 +2,6 @@
 
 pub mod execution_protocol;
 pub mod scheduler;
-pub mod scheduler_protocol;
 pub mod work_item;
 pub mod workspace;
