--- tests/support/scheduler_workitem_mvp.rs
@@ -1,1 +1,0 @@
-pub use holon::domain::scheduler_protocol::*;
